OrmLite UpdateUserAuth check for null password on validation
Added a simple check to allow the password to be null during the
UpdateUserAuth method.  This way you can update a user without having to
update the password as well.  All the validation pointed the ability of
passing in a null password, so I pass in "this-is-a-hack" string into
the ValidateNewUser method.